feat: add admin rating and user management APIs
- Implement admin APIs for managing product ratings.
- Add endpoints to get, approve, delete, and ban ratings.
- Create user management APIs to list users and get user info.
- Introduce pagination for user listing.
- Add task distribution for indexing products in Elasticsearch.
- Remove obsolete Elasticsearch processing code.
- Create new task messages table for managing task statuses.
