+++ internal/api/admin_product.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"encoding/json"
 	"errors"
 	"net/http"
 	"slices"
@@ -8,13 +9,13 @@ import (
 	"strings"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/go-chi/jwtauth/v5"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
 	"github.com/thanhphuocnguyen/go-eshop/internal/db/repository"
 	"github.com/thanhphuocnguyen/go-eshop/internal/dto"
 	"github.com/thanhphuocnguyen/go-eshop/internal/models"
 	"github.com/thanhphuocnguyen/go-eshop/internal/utils"
+	"github.com/thanhphuocnguyen/go-eshop/internal/worker"
 )
 
 // Setup admin-related routes
@@ -134,96 +135,6 @@ func (s *Server) addAdminRoutes(r chi.Router) {
 	})
 }
 
-// adminGetUsers godoc
-// @Summary List users
-// @Description List users
-// @Tags users
-// @Accept  json
-// @Produce  json
-// @Param limit query int false "Limit"
-// @Param offset query int false "Offset"
-// @Success 200 {object} dto.ApiResponse[[]dto.UserDetail]
-// @Failure 500 {object} ErrorResp
-// @Failure 400 {object} ErrorResp
-// @Failure 401 {object} ErrorResp
-// @Router /admin/users [get]
-func (s *Server) adminGetUsers(w http.ResponseWriter, r *http.Request) {
-	c := r.Context()
-	_, claims, err := jwtauth.FromContext(c)
-	if err != nil {
-		RespondInternalServerError(w, InternalServerErrorCode, errors.New("authorization payload is not provided"))
-		return
-	}
-
-	// Parse query parameters
-	queries := ParsePaginationQuery(r)
-
-	users, err := s.repo.GetUsers(c, repository.GetUsersParams{
-		Limit:  queries.PageSize,
-		Offset: (queries.Page - 1) * queries.PageSize,
-	})
-
-	if err != nil {
-		RespondInternalServerError(w, InternalServerErrorCode, err)
-		return
-	}
-
-	total, err := s.repo.CountUsers(c)
-	if err != nil {
-		RespondInternalServerError(w, InternalServerErrorCode, err)
-		return
-	}
-
-	userResp := make([]dto.UserDetail, 0)
-	roleCode := claims["role_code"].(string)
-	for _, user := range users {
-		userResp = append(userResp, dto.MapToUserResponse(user, roleCode))
-	}
-
-	pagination := dto.CreatePagination(queries.Page, queries.PageSize, total)
-	RespondSuccessWithPagination(w, userResp, pagination)
-}
-
-// adminGetUser godoc
-// @Summary Get user info
-// @Description Get user info
-// @Tags admin
-// @Accept  json
-// @Produce  json
-// @Param id path string true "User ID"
-// @Success 200 {object} dto.ApiResponse[dto.UserDetail]
-// @Failure 400 {object} ErrorResp
-// @Failure 404 {object} ErrorResp
-// @Failure 500 {object} ErrorResp
-// @Router /admin/users/{id} [get]
-func (s *Server) adminGetUser(w http.ResponseWriter, r *http.Request) {
-	c := r.Context()
-	_, claims, err := jwtauth.FromContext(c)
-	if err != nil {
-		RespondInternalServerError(w, InternalServerErrorCode, errors.New("authorization payload is not provided"))
-		return
-	}
-
-	id, err := GetUrlParam(r, "id")
-	if err != nil {
-		RespondBadRequest(w, InvalidBodyCode, err)
-		return
-	}
-
-	user, err := s.repo.GetUserByID(c, uuid.MustParse(id))
-	if err != nil {
-		if errors.Is(err, repository.ErrRecordNotFound) {
-			RespondNotFound(w, NotFoundCode, err)
-			return
-		}
-		RespondInternalServerError(w, InternalServerErrorCode, err)
-		return
-	}
-	roleCode := claims["role_code"].(string)
-	userResp := dto.MapToUserResponse(user, roleCode)
-	RespondSuccess(w, userResp)
-}
-
 // adminGetProducts godoc
 // @Summary Get admin list of products
 // @Schemes http
@@ -338,6 +249,12 @@ func (s *Server) adminAddProduct(w http.ResponseWriter, r *http.Request) {
 		RespondInternalServerError(w, InternalServerErrorCode, err)
 		return
 	}
+	err = s.taskDistributor.SendIndexProductTask(c, &worker.PayloadIndexProduct{
+		ProductID: product.ID,
+	})
+	if err != nil {
+		log.Error().Err(err).Msg("SendIndexProductTask")
+	}
 
 	RespondCreated(w, product)
 }
@@ -424,6 +341,13 @@ func (s *Server) adminUpdateProduct(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	err = s.taskDistributor.SendUpdateIndexProductTask(c, &worker.PayloadIndexProduct{
+		ProductID: updated.ID,
+	})
+	if err != nil {
+		log.Error().Err(err).Msg("SendUpdateIndexProductTask")
+	}
+
 	RespondSuccess(w, updated)
 }
 
@@ -461,7 +385,12 @@ func (s *Server) adminDeleteProduct(w http.ResponseWriter, r *http.Request) {
 		RespondInternalServerError(w, InternalServerErrorCode, err)
 		return
 	}
-
+	err = s.taskDistributor.SendDeleteIndexProductTask(c, &worker.PayloadIndexProduct{
+		ProductID: product.ID,
+	})
+	if err != nil {
+		log.Error().Err(err).Msg("SendDeleteIndexProductTask")
+	}
 	RespondNoContent(w)
 }
 
@@ -638,6 +567,13 @@ func (s *Server) adminAddVariant(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	err = s.taskDistributor.SendIndexProductTask(c, &worker.PayloadIndexProduct{
+		ProductID: prod.ID,
+	})
+	if err != nil {
+		log.Error().Err(err).Msg("SendIndexProductTask")
+	}
+
 	RespondCreated(w, variant)
 }
 
@@ -745,4 +681,197 @@ func (s *Server) getVariantByProductId(w http.ResponseWriter, r *http.Request) {
 	}
 
 	RespondSuccess(w, resp)
+}
+
+// @Summary Update a product variant
+// @Schemes http
+// @Description update a product with the input payload
+// @Tags products
+// @Accept json
+// @Param input body models.UpdateProdVariantModel true "Product variant input"
+// @Produce json
+// @Success 200 {object} dto.ApiResponse[repository.ProductVariant]
+// @Failure 400 {object} ErrorResp
+// @Failure 500 {object} ErrorResp
+// @Router /admin/products/{id}/variants/{variantId} [put]
+func (s *Server) adminUpdateVariant(w http.ResponseWriter, r *http.Request) {
+	c := r.Context()
+	id := chi.URLParam(r, "id")
+	variantId := chi.URLParam(r, "variantId")
+	if id == "" || variantId == "" {
+		RespondBadRequest(w, InvalidBodyCode, errors.New("id and variantId parameters are required"))
+		return
+	}
+
+	var req models.UpdateProdVariantModel
+	if err := s.GetRequestBody(r, &req); err != nil {
+		RespondBadRequest(w, InvalidBodyCode, err)
+		return
+	}
+
+	prod, err := s.repo.GetProductByID(c, repository.GetProductByIDParams{ID: uuid.MustParse(id)})
+	if err != nil {
+		RespondNotFound(w, NotFoundCode, err)
+		return
+	}
+
+	updateParams := repository.UpdateProductVariantParams{
+		ProductID: prod.ID,
+		ID:        uuid.MustParse(variantId),
+	}
+
+	if req.Price != nil {
+		updateParams.Price = utils.GetPgNumericFromFloat(*req.Price)
+	}
+	if req.StockQty != nil {
+		updateParams.Stock = req.StockQty
+	}
+	if req.Weight != nil {
+		updateParams.Weight = utils.GetPgNumericFromFloat(*req.Weight)
+	}
+	if req.Description != nil {
+		updateParams.Description = req.Description
+	}
+
+	updatedVariant, err := s.repo.UpdateProductVariant(c, updateParams)
+	if err != nil {
+		log.Error().Err(err).Timestamp().Msg("UpdateProductVariant")
+		RespondInternalServerError(w, InternalServerErrorCode, err)
+		return
+	}
+
+	RespondSuccess(w, updatedVariant)
+}
+
+// @Summary Upload a product variant image
+// @Schemes http
+// @Description upload a product variant image with the input payload
+// @Tags products
+// @Accept multipart/form-data
+// @Param id path string true "Product ID"
+// @Param variantId path string true "Product Variant ID"
+// @Produce json
+// @Success 200 {object} dto.ApiResponse[repository.ProductVariant]
+// @Failure 400 {object} ErrorResp
+// @Failure 500 {object} ErrorResp
+// @Router /admin/products/{id}/variants/{variantId}/images [post]
+func (s *Server) adminUploadVariantImage(w http.ResponseWriter, r *http.Request) {
+	c := r.Context()
+	id := chi.URLParam(r, "id")
+	variantId := chi.URLParam(r, "variantId")
+	if id == "" || variantId == "" {
+		RespondBadRequest(w, InvalidBodyCode, errors.New("id and variantId parameters are required"))
+		return
+	}
+
+	// Parse multipart form
+	err := r.ParseMultipartForm(10 << 20) // 10MB max
+	if err != nil {
+		RespondBadRequest(w, InvalidBodyCode, err)
+		return
+	}
+
+	file, header, err := r.FormFile("file")
+	if err != nil {
+		RespondBadRequest(w, InvalidBodyCode, err)
+		return
+	}
+	defer file.Close()
+
+	prod, err := s.repo.GetProductByID(c, repository.GetProductByIDParams{ID: uuid.MustParse(id)})
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(w).Encode(dto.CreateErr(NotFoundCode, err))
+		return
+	}
+
+	variant, err := s.repo.GetProductVariantByID(c, repository.GetProductVariantByIDParams{
+		ID:        uuid.MustParse(variantId),
+		ProductID: prod.ID,
+	})
+
+	if err != nil {
+		RespondNotFound(w, NotFoundCode, err)
+		return
+	}
+
+	if variant.ImageID != nil {
+		msg, err := s.uploadService.Remove(c, *variant.ImageID)
+		if err != nil {
+			log.Error().Err(err).Timestamp().Msg(msg)
+			RespondInternalServerError(w, InternalServerErrorCode, err)
+			return
+		}
+	}
+
+	fileHeader := &struct {
+		Filename string
+		Header   map[string][]string
+		Size     int64
+	}{
+		Filename: header.Filename,
+		Header:   header.Header,
+		Size:     header.Size,
+	}
+
+	uploadID, url, err := s.uploadService.Upload(c, fileHeader)
+	if err != nil {
+		log.Error().Err(err).Timestamp().Msg("UploadFile")
+		RespondInternalServerError(w, InternalServerErrorCode, err)
+		return
+	}
+	updateParam := repository.UpdateProductVariantParams{
+		ProductID: prod.ID,
+		ID:        variant.ID,
+		ImageID:   &uploadID,
+		ImageUrl:  &url,
+	}
+
+	updatedVariant, err := s.repo.UpdateProductVariant(c, updateParam)
+	if err != nil {
+		log.Error().Err(err).Timestamp().Msg("UpdateProductVariant")
+		RespondInternalServerError(w, InternalServerErrorCode, err)
+		return
+	}
+
+	RespondSuccess(w, updatedVariant)
+}
+
+// @Summary Delete a product variant
+// @Schemes http
+// @Description delete a product variant with the input payload
+// @Tags products
+// @Accept json
+// @Produce json
+// @Success 200 {object} dto.ApiResponse[string]
+// @Failure 400 {object} ErrorResp
+// @Failure 500 {object} ErrorResp
+// @Router /admin/products/{id}/variant/{variantID} [delete]
+func (s *Server) adminDeleteVariant(w http.ResponseWriter, r *http.Request) {
+	c := r.Context()
+	id := chi.URLParam(r, "id")
+	variantId := chi.URLParam(r, "variantId")
+	if id == "" || variantId == "" {
+		RespondBadRequest(w, InvalidBodyCode, errors.New("id and variantId parameters are required"))
+		return
+	}
+
+	prod, err := s.repo.GetProductByID(c, repository.GetProductByIDParams{ID: uuid.MustParse(id)})
+	if err != nil {
+		RespondNotFound(w, NotFoundCode, err)
+		return
+	}
+
+	err = s.repo.DeleteProductVariant(c, repository.DeleteProductVariantParams{
+		ProductID: prod.ID,
+		ID:        uuid.MustParse(variantId),
+	})
+
+	if err != nil {
+		log.Error().Err(err).Timestamp().Msg("DeleteVariant")
+		RespondInternalServerError(w, InternalServerErrorCode, err)
+		return
+	}
+	RespondNoContent(w)
 }
