+++ internal/worker/task_names.go
@@ -1,7 +1,9 @@
 package worker
 
 const (
-	OrderCreatedEmailTaskType = "order_created_email"
-	VerifyEmailTaskType       = "send_verify_email"
-	IndexProductTaskType      = "index_product"
+	OrderCreatedEmailTaskType  = "order_created_email"
+	VerifyEmailTaskType        = "send_verify_email"
+	IndexProductTaskType       = "index_product"
+	UpdateIndexProductTaskType = "update_index_product"
+	DeleteIndexProductTaskType = "delete_index_product"
 )
